Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
certifi==2025.6.15
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.13
decorator==5.1.1
distro==1.9.0
Django==4.2.21
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jiter==0.9.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.70.0
packaging==24.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.3.7
prompt-toolkit==3.0.43
psutil==7.0.0
pure-eval==0.2.2
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.17.2
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.3.post1
pywin32==310
pyzmq==26.3.0
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.3
six==1.15.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
zipp==3.21.0
drf-yasg==1.21.7
